+++ lib/screen/wallet_connect/v2/wc2_permission_page.dart
@@ -620,4 +620,3 @@ class _PersonalConnectItemState extends State<PersonalConnectItem> {
     );
   }
 }
-
